メニュー型定義を分離 (TypeScriptの型支援が効かないので)
